test: rewrite SSLConfig race test as subprocess + fixture
Subprocess isolation gives cleaner crash detection: if the race triggers
in the fixture, the subprocess exits non-zero and the test fails via
expect(exitCode).toBe(0). On fixed code, subprocess exits 0.

Fixture still uses workers internally (required since GlobalRegistry is
per-process — the race needs independent JS threads calling intern()
while the HTTP thread derefs).

Subprocess gets BUN_DEBUG_SSLConfig=1 explicitly (bunEnv strips
BUN_DEBUG_* vars). In debug builds this enables scoped logging that
widens the race window. No-op in release builds.
